+++ packages/base/config.js
@@ -1,4 +1,5 @@
 import js from "@eslint/js";
+import { defineConfig } from "eslint/config";
 import prettier from "eslint-config-prettier";
 import { createTypeScriptImportResolver } from "eslint-import-resolver-typescript";
 import importPlugin from "eslint-plugin-import-x";
@@ -7,7 +8,7 @@ import importSortPlugin from "eslint-plugin-simple-import-sort";
 import globals from "globals";
 import ts from "typescript-eslint";
 
-const config = ts.config(
+const config = defineConfig(
 	js.configs.recommended,
 	ts.configs.strictTypeChecked,
 	ts.configs.stylisticTypeChecked,