+++ packages/mdx/config.js
@@ -1,16 +1,16 @@
+import { defineConfig } from "eslint/config";
 import * as mdx from "eslint-plugin-mdx";
-import ts from "typescript-eslint";
 
-const config = ts.config(
+const config = defineConfig(
 	{
 		name: "acdh-oeaw/mdx-config",
-		extends: mdx.flat,
+		extends: [mdx.flat],
 		/** Lint code blocks. */
 		processor: mdx.createRemarkProcessor({ lintCodeBlocks: true }),
 	},
 	{
 		name: "acdh-oeaw/mdx-config/code-blocks",
-		extends: mdx.flatCodeBlocks,
+		extends: [mdx.flatCodeBlocks],
 	},
 );