+++ packages/playwright/config.js
@@ -1,10 +1,10 @@
+import { defineConfig } from "eslint/config";
 import playwrightPlugin from "eslint-plugin-playwright";
-import ts from "typescript-eslint";
 
-const config = ts.config({
+const config = defineConfig({
 	name: "acdh-oeaw/playwright-config",
 	files: ["e2e/**/*.@(spec|test).@(ts|tsx)"],
-	extends: playwrightPlugin.configs["flat/recommended"],
+	extends: [playwrightPlugin.configs["flat/recommended"]],
 });
 
 export default config;